@@ -29,7 +29,7 @@ authlib = "0.15.5"
 backoff = "2.2.1"
 blinker = "1.6.2"
 click = "8.1.3"
-cryptography = "41.0.6"
+cryptography = "43.0.1"
 disposable-email-domains = ">=0.0.52"
 flask = "2.3.2"
 flask-limiter = "3.3.1"
@@ -43,10 +43,10 @@ flask-wtf = "1.1.1"
 funcy = "1.13"
 gevent = "23.9.1"
 greenlet = "2.0.2"
-gunicorn = "20.0.4"
+gunicorn = "22.0.0"
 httplib2 = "0.19.0"
 itsdangerous = "2.1.2"
-jinja2 = "3.1.3"
+jinja2 = "3.1.5"
 jsonschema = "3.1.1"
 markupsafe = "2.1.1"
 maxminddb-geolite2 = "2018.703"
@@ -54,7 +54,7 @@ parsedatetime = "2.4"
 passlib = "1.7.3"
 psycopg2-binary = "2.9.6"
 pyjwt = "2.4.0"
-pyopenssl = "23.2.0"
+pyopenssl = "24.2.1"
 pypd = "1.1.0"
 pysaml2 = "7.3.1"
 pystache = "0.6.0"
@@ -64,34 +64,38 @@ pytz = ">=2019.3"
 pyyaml = "6.0.1"
 redis = "4.6.0"
 regex = "2023.8.8"
-requests = "2.31.0"
-restrictedpython = "6.2"
-rq = "1.9.0"
-rq-scheduler = "0.11.0"
+requests = "2.32.3"
+restrictedpython = "7.3"
+rq = "1.16.1"
+rq-scheduler = "0.13.1"
 semver = "2.8.1"
-sentry-sdk = "1.28.1"
+sentry-sdk = "1.45.1"
 sqlalchemy = "1.3.24"
 sqlalchemy-searchable = "1.2.0"
-sqlalchemy-utils = "0.34.2"
-sqlparse = "0.4.4"
+sqlalchemy-utils = "0.38.3"
+sqlparse = "0.5.0"
 sshtunnel = "0.1.5"
 statsd = "3.3.0"
 supervisor = "4.1.0"
 supervisor-checks = "0.8.1"
 ua-parser = "0.18.0"
-urllib3 = "1.26.18"
+urllib3 = "1.26.19"
 user-agents = "2.0"
 werkzeug = "2.3.8"
 wtforms = "2.2.1"
 xlsxwriter = "1.2.2"
 tzlocal = "4.3.1"
+pyodbc = "5.1.0"
+debugpy = "^1.8.9"
+paramiko = "3.4.1"
+oracledb = "2.5.1"
 
 [tool.poetry.group.all_ds]
 optional = true
 
 [tool.poetry.group.all_ds.dependencies]
 atsd-client = "3.0.5"
-azure-kusto-data = "0.0.35"
+azure-kusto-data = "5.0.1"
 boto3 = "1.28.8"
 botocore = "1.31.8"
 cassandra-driver = "3.21.0"
@@ -106,30 +110,30 @@ influxdb = "5.2.3"
 influxdb-client = "1.38.0"
 memsql = "3.2.0"
 mysqlclient = "2.1.1"
+numpy = "1.24.4"
 nzalchemy = "^11.0.2"
 nzpy = ">=1.15"
 oauth2client = "4.1.3"
 openpyxl = "3.0.7"
-oracledb = "2.0.1"
 pandas = "1.3.4"
 phoenixdb = "0.7"
 pinotdb = ">=0.4.5"
 protobuf = "3.20.2"
-pyathena = ">=1.5.0,<=1.11.5"
+pyathena = "2.25.2"
 pydgraph = "2.0.2"
 pydruid = "0.5.7"
 pyexasol = "0.12.0"
 pyhive = "0.6.1"
 pyignite = "0.6.1"
-pymongo = { version = "4.3.3", extras = ["srv", "tls"] }
-pymssql = "2.2.8"
-pyodbc = "4.0.28"
+pymongo = { version = "4.6.3", extras = ["srv", "tls"] }
+pymssql = "^2.3.1"
+pyodbc = "5.1.0"
 python-arango = "6.1.0"
-python-rapidjson = "1.1.0"
+python-rapidjson = "1.20"
 requests-aws-sign = "0.1.5"
 sasl = ">=0.1.3"
 simple-salesforce = "0.74.3"
-snowflake-connector-python = "3.4.0"
+snowflake-connector-python = "3.12.3"
 td-client = "1.0.0"
 thrift = ">=0.8.0"
 thrift-sasl = ">=0.1.0"
@@ -155,7 +159,6 @@ jwcrypto = "1.5.6"
 mock = "5.0.2"
 pre-commit = "3.3.3"
 ptpython = "3.0.23"
-ptvsd = "4.3.2"
 pytest-cov = "4.1.0"
 watchdog = "3.0.0"
 ruff = "0.0.289"
