@@ -44,7 +44,7 @@ export const Columns = {
   date(overrides) {
     return extend(
       {
-        render: text => formatDate(text),
+        render: (text) => formatDate(text),
       },
       overrides
     );
@@ -52,7 +52,7 @@ export const Columns = {
   dateTime(overrides) {
     return extend(
       {
-        render: text => formatDateTime(text),
+        render: (text) => formatDateTime(text),
       },
       overrides
     );
@@ -62,7 +62,7 @@ export const Columns = {
       {
         width: "1%",
         className: "text-nowrap",
-        render: text => durationHumanize(text),
+        render: (text) => durationHumanize(text),
       },
       overrides
     );
@@ -70,7 +70,7 @@ export const Columns = {
   timeAgo(overrides, timeAgoCustomProps = undefined) {
     return extend(
       {
-        render: value => <TimeAgo date={value} {...timeAgoCustomProps} />,
+        render: (value) => <TimeAgo date={value} {...timeAgoCustomProps} />,
       },
       overrides
     );
@@ -110,6 +110,7 @@ export default class ItemsTable extends React.Component {
     orderByField: PropTypes.string,
     orderByReverse: PropTypes.bool,
     toggleSorting: PropTypes.func,
+    setSorting: PropTypes.func,
     "data-test": PropTypes.string,
     rowKey: PropTypes.oneOfType([PropTypes.string, PropTypes.func]),
   };
@@ -127,18 +128,15 @@ export default class ItemsTable extends React.Component {
   };
 
   prepareColumns() {
-    const { orderByField, orderByReverse, toggleSorting } = this.props;
+    const { orderByField, orderByReverse } = this.props;
     const orderByDirection = orderByReverse ? "descend" : "ascend";
 
     return map(
       map(
-        filter(this.props.columns, column => (isFunction(column.isAvailable) ? column.isAvailable() : true)),
-        column => extend(column, { orderByField: column.orderByField || column.field })
+        filter(this.props.columns, (column) => (isFunction(column.isAvailable) ? column.isAvailable() : true)),
+        (column) => extend(column, { orderByField: column.orderByField || column.field })
       ),
       (column, index) => {
-        // Bind click events only to sortable columns
-        const onHeaderCell = column.sorter ? () => ({ onClick: () => toggleSorting(column.orderByField) }) : null;
-
         // Wrap render function to pass correct arguments
         const render = isFunction(column.render) ? (text, row) => column.render(text, row.item) : identity;
 
@@ -146,14 +144,13 @@ export default class ItemsTable extends React.Component {
           key: "column" + index,
           dataIndex: ["item", column.field],
           defaultSortOrder: column.orderByField === orderByField ? orderByDirection : null,
-          onHeaderCell,
           render,
         });
       }
     );
   }
 
-  getRowKey = record => {
+  getRowKey = (record) => {
     const { rowKey } = this.props;
     if (rowKey) {
       if (isFunction(rowKey)) {
@@ -172,22 +169,43 @@ export default class ItemsTable extends React.Component {
 
     // Bind events only if `onRowClick` specified
     const onTableRow = isFunction(this.props.onRowClick)
-      ? row => ({
-          onClick: event => {
+      ? (row) => ({
+          onClick: (event) => {
             this.props.onRowClick(event, row.item);
           },
         })
       : null;
 
+    const onChange = (pagination, filters, sorter, extra) => {
+      const action = extra?.action;
+      if (action === "sort") {
+        const propsColumn = this.props.columns.find((column) => column.field === sorter.field[1]);
+        if (!propsColumn.sorter) {
+          return;
+        }
+        let orderByField = propsColumn.orderByField;
+        const orderByReverse = sorter.order === "descend";
+
+        if (orderByReverse === undefined) {
+          orderByField = null;
+        }
+        if (this.props.setSorting) {
+          this.props.setSorting(orderByField, orderByReverse);
+        } else {
+          this.props.toggleSorting(orderByField);
+        }
+      }
+    };
+
     const { showHeader } = this.props;
     if (this.props.loading) {
       if (isEmpty(tableDataProps.dataSource)) {
-        tableDataProps.columns = tableDataProps.columns.map(column => ({
+        tableDataProps.columns = tableDataProps.columns.map((column) => ({
           ...column,
           sorter: false,
           render: () => <Skeleton active paragraph={false} />,
         }));
-        tableDataProps.dataSource = range(10).map(key => ({ key: `${key}` }));
+        tableDataProps.dataSource = range(10).map((key) => ({ key: `${key}` }));
       } else {
         tableDataProps.loading = { indicator: null };
       }
@@ -200,6 +218,7 @@ export default class ItemsTable extends React.Component {
         rowKey={this.getRowKey}
         pagination={false}
         onRow={onTableRow}
+        onChange={onChange}
         data-test={this.props["data-test"]}
         {...tableDataProps}
       />
