@@ -61,7 +61,7 @@ class DateParameter extends Parameter {
       return value;
     }
 
-    const normalizedValue = moment(value);
+    const normalizedValue = moment(value, moment.ISO_8601, true);
     return normalizedValue.isValid() ? normalizedValue : null;
   }
 
