@@ -2,16 +2,14 @@ import { dragParam } from "../../support/parameters";
 import dayjs from "dayjs";
 
 function openAndSearchAntdDropdown(testId, paramOption) {
-  cy.getByTestId(testId)
-    .find(".ant-select-selection-search-input")
-    .type(paramOption, { force: true });
+  cy.getByTestId(testId).find(".ant-select-selection-search-input").type(paramOption, { force: true });
 }
 
 describe("Parameter", () => {
-  const expectDirtyStateChange = edit => {
+  const expectDirtyStateChange = (edit) => {
     cy.getByTestId("ParameterName-test-parameter")
       .find(".parameter-input")
-      .should($el => {
+      .should(($el) => {
         assert.isUndefined($el.data("dirty"));
       });
 
@@ -19,7 +17,7 @@ describe("Parameter", () => {
 
     cy.getByTestId("ParameterName-test-parameter")
       .find(".parameter-input")
-      .should($el => {
+      .should(($el) => {
         assert.isTrue($el.data("dirty"));
       });
   };
@@ -42,9 +40,7 @@ describe("Parameter", () => {
     });
 
     it("updates the results after clicking Apply", () => {
-      cy.getByTestId("ParameterName-test-parameter")
-        .find("input")
-        .type("Redash");
+      cy.getByTestId("ParameterName-test-parameter").find("input").type("Redash");
 
       cy.getByTestId("ParameterApplyButton").click();
 
@@ -53,10 +49,63 @@ describe("Parameter", () => {
 
     it("sets dirty state when edited", () => {
       expectDirtyStateChange(() => {
-        cy.getByTestId("ParameterName-test-parameter")
-          .find("input")
-          .type("Redash");
+        cy.getByTestId("ParameterName-test-parameter").find("input").type("Redash");
       });
+    });
+  });
+
+  describe("Text Pattern Parameter", () => {
+    beforeEach(() => {
+      const queryData = {
+        name: "Text Pattern Parameter",
+        query: "SELECT '{{test-parameter}}' AS parameter",
+        options: {
+          parameters: [{ name: "test-parameter", title: "Test Parameter", type: "text-pattern", regex: "a.*a" }],
+        },
+      };
+
+      cy.createQuery(queryData, false).then(({ id }) => cy.visit(`/queries/${id}/source`));
+    });
+
+    it("updates the results after clicking Apply", () => {
+      cy.getByTestId("ParameterName-test-parameter").find("input").type("{selectall}arta");
+
+      cy.getByTestId("ParameterApplyButton").click();
+
+      cy.getByTestId("TableVisualization").should("contain", "arta");
+
+      cy.getByTestId("ParameterName-test-parameter").find("input").type("{selectall}arounda");
+
+      cy.getByTestId("ParameterApplyButton").click();
+
+      cy.getByTestId("TableVisualization").should("contain", "arounda");
+    });
+
+    it("throws error message with invalid query request", () => {
+      cy.getByTestId("ParameterName-test-parameter").find("input").type("{selectall}arta");
+
+      cy.getByTestId("ParameterApplyButton").click();
+
+      cy.getByTestId("ParameterName-test-parameter").find("input").type("{selectall}abcab");
+
+      cy.getByTestId("ParameterApplyButton").click();
+
+      cy.getByTestId("QueryExecutionStatus").should("exist");
+    });
+
+    it("sets dirty state when edited", () => {
+      expectDirtyStateChange(() => {
+        cy.getByTestId("ParameterName-test-parameter").find("input").type("{selectall}arta");
+      });
+    });
+
+    it("doesn't let user save invalid regex", () => {
+      cy.get(".fa-cog").click();
+      cy.getByTestId("RegexPatternInput").type("{selectall}[");
+      cy.contains("Invalid Regex Pattern").should("exist");
+      cy.getByTestId("SaveParameterSettings").click();
+      cy.get(".fa-cog").click();
+      cy.getByTestId("RegexPatternInput").should("not.equal", "[");
     });
   });
 
@@ -74,17 +123,13 @@ describe("Parameter", () => {
     });
 
     it("updates the results after clicking Apply", () => {
-      cy.getByTestId("ParameterName-test-parameter")
-        .find("input")
-        .type("{selectall}42");
+      cy.getByTestId("ParameterName-test-parameter").find("input").type("{selectall}42");
 
       cy.getByTestId("ParameterApplyButton").click();
 
       cy.getByTestId("TableVisualization").should("contain", 42);
 
-      cy.getByTestId("ParameterName-test-parameter")
-        .find("input")
-        .type("{selectall}31415");
+      cy.getByTestId("ParameterName-test-parameter").find("input").type("{selectall}31415");
 
       cy.getByTestId("ParameterApplyButton").click();
 
@@ -93,9 +138,7 @@ describe("Parameter", () => {
 
     it("sets dirty state when edited", () => {
       expectDirtyStateChange(() => {
-        cy.getByTestId("ParameterName-test-parameter")
-          .find("input")
-          .type("{selectall}42");
+        cy.getByTestId("ParameterName-test-parameter").find("input").type("{selectall}42");
       });
     });
   });
@@ -119,10 +162,7 @@ describe("Parameter", () => {
       openAndSearchAntdDropdown("ParameterName-test-parameter", "value2"); // asserts option filter prop
 
       // only the filtered option should be on the DOM
-      cy.get(".ant-select-item-option")
-        .should("have.length", 1)
-        .and("contain", "value2")
-        .click();
+      cy.get(".ant-select-item-option").should("have.length", 1).and("contain", "value2").click();
 
       cy.getByTestId("ParameterApplyButton").click();
       // ensure that query is being executed
@@ -140,12 +180,10 @@ describe("Parameter", () => {
         SaveParameterSettings
       `);
 
-      cy.getByTestId("ParameterName-test-parameter")
-        .find(".ant-select-selection-search")
-        .click();
+      cy.getByTestId("ParameterName-test-parameter").find(".ant-select-selection-search").click();
 
       // select all unselected options
-      cy.get(".ant-select-item-option").each($option => {
+      cy.get(".ant-select-item-option").each(($option) => {
         if (!$option.hasClass("ant-select-item-option-selected")) {
           cy.wrap($option).click();
         }
@@ -160,9 +198,7 @@ describe("Parameter", () => {
 
     it("sets dirty state when edited", () => {
       expectDirtyStateChange(() => {
-        cy.getByTestId("ParameterName-test-parameter")
-          .find(".ant-select")
-          .click();
+        cy.getByTestId("ParameterName-test-parameter").find(".ant-select").click();
 
         cy.contains(".ant-select-item-option", "value2").click();
       });
@@ -176,7 +212,7 @@ describe("Parameter", () => {
           name: "Dropdown Query",
           query: "",
         };
-        cy.createQuery(dropdownQueryData, true).then(dropdownQuery => {
+        cy.createQuery(dropdownQueryData, true).then((dropdownQuery) => {
           const queryData = {
             name: "Query Based Dropdown Parameter",
             query: "SELECT '{{test-parameter}}' AS parameter",
@@ -208,7 +244,7 @@ describe("Parameter", () => {
                   SELECT 'value2' AS name, 2 AS value UNION ALL
                   SELECT 'value3' AS name, 3 AS value`,
         };
-        cy.createQuery(dropdownQueryData, true).then(dropdownQuery => {
+        cy.createQuery(dropdownQueryData, true).then((dropdownQuery) => {
           const queryData = {
             name: "Query Based Dropdown Parameter",
             query: "SELECT '{{test-parameter}}' AS parameter",
@@ -234,10 +270,7 @@ describe("Parameter", () => {
         openAndSearchAntdDropdown("ParameterName-test-parameter", "value2"); // asserts option filter prop
 
         // only the filtered option should be on the DOM
-        cy.get(".ant-select-item-option")
-          .should("have.length", 1)
-          .and("contain", "value2")
-          .click();
+        cy.get(".ant-select-item-option").should("have.length", 1).and("contain", "value2").click();
 
         cy.getByTestId("ParameterApplyButton").click();
         // ensure that query is being executed
@@ -255,12 +288,10 @@ describe("Parameter", () => {
           SaveParameterSettings
         `);
 
-        cy.getByTestId("ParameterName-test-parameter")
-          .find(".ant-select")
-          .click();
+        cy.getByTestId("ParameterName-test-parameter").find(".ant-select").click();
 
         // make sure all options are unselected and select all
-        cy.get(".ant-select-item-option").each($option => {
+        cy.get(".ant-select-item-option").each(($option) => {
           expect($option).not.to.have.class("ant-select-dropdown-menu-item-selected");
           cy.wrap($option).click();
         });
@@ -274,14 +305,10 @@ describe("Parameter", () => {
     });
   });
 
-  const selectCalendarDate = date => {
-    cy.getByTestId("ParameterName-test-parameter")
-      .find("input")
-      .click();
+  const selectCalendarDate = (date) => {
+    cy.getByTestId("ParameterName-test-parameter").find("input").click();
 
-    cy.get(".ant-picker-panel")
-      .contains(".ant-picker-cell-inner", date)
-      .click();
+    cy.get(".ant-picker-panel").contains(".ant-picker-cell-inner", date).click();
   };
 
   describe("Date Parameter", () => {
@@ -303,10 +330,10 @@ describe("Parameter", () => {
     });
 
     afterEach(() => {
-      cy.clock().then(clock => clock.restore());
+      cy.clock().then((clock) => clock.restore());
     });
 
-    it("updates the results after selecting a date", function() {
+    it("updates the results after selecting a date", function () {
       selectCalendarDate("15");
 
       cy.getByTestId("ParameterApplyButton").click();
@@ -314,12 +341,10 @@ describe("Parameter", () => {
       cy.getByTestId("TableVisualization").should("contain", dayjs(this.now).format("15/MM/YY"));
     });
 
-    it("allows picking a dynamic date", function() {
+    it("allows picking a dynamic date", function () {
       cy.getByTestId("DynamicButton").click();
 
-      cy.getByTestId("DynamicButtonMenu")
-        .contains("Today/Now")
-        .click();
+      cy.getByTestId("DynamicButtonMenu").contains("Today/Now").click();
 
       cy.getByTestId("ParameterApplyButton").click();
 
@@ -350,14 +375,11 @@ describe("Parameter", () => {
     });
 
     afterEach(() => {
-      cy.clock().then(clock => clock.restore());
+      cy.clock().then((clock) => clock.restore());
     });
 
-    it("updates the results after selecting a date and clicking in ok", function() {
-      cy.getByTestId("ParameterName-test-parameter")
-        .find("input")
-        .as("Input")
-        .click();
+    it("updates the results after selecting a date and clicking in ok", function () {
+      cy.getByTestId("ParameterName-test-parameter").find("input").as("Input").click();
 
       selectCalendarDate("15");
 
@@ -368,27 +390,20 @@ describe("Parameter", () => {
       cy.getByTestId("TableVisualization").should("contain", dayjs(this.now).format("YYYY-MM-15 HH:mm"));
     });
 
-    it("shows the current datetime after clicking in Now", function() {
-      cy.getByTestId("ParameterName-test-parameter")
-        .find("input")
-        .as("Input")
-        .click();
+    it("shows the current datetime after clicking in Now", function () {
+      cy.getByTestId("ParameterName-test-parameter").find("input").as("Input").click();
 
-      cy.get(".ant-picker-panel")
-        .contains("Now")
-        .click();
+      cy.get(".ant-picker-panel").contains("Now").click();
 
       cy.getByTestId("ParameterApplyButton").click();
 
       cy.getByTestId("TableVisualization").should("contain", dayjs(this.now).format("YYYY-MM-DD HH:mm"));
     });
 
-    it("allows picking a dynamic date", function() {
+    it("allows picking a dynamic date", function () {
       cy.getByTestId("DynamicButton").click();
 
-      cy.getByTestId("DynamicButtonMenu")
-        .contains("Today/Now")
-        .click();
+      cy.getByTestId("DynamicButtonMenu").contains("Today/Now").click();
 
       cy.getByTestId("ParameterApplyButton").click();
 
@@ -397,31 +412,20 @@ describe("Parameter", () => {
 
     it("sets dirty state when edited", () => {
       expectDirtyStateChange(() => {
-        cy.getByTestId("ParameterName-test-parameter")
-          .find("input")
-          .click();
+        cy.getByTestId("ParameterName-test-parameter").find("input").click();
 
-        cy.get(".ant-picker-panel")
-          .contains("Now")
-          .click();
+        cy.get(".ant-picker-panel").contains("Now").click();
       });
     });
   });
 
   describe("Date Range Parameter", () => {
     const selectCalendarDateRange = (startDate, endDate) => {
-      cy.getByTestId("ParameterName-test-parameter")
-        .find("input")
-        .first()
-        .click();
+      cy.getByTestId("ParameterName-test-parameter").find("input").first().click();
 
-      cy.get(".ant-picker-panel")
-        .contains(".ant-picker-cell-inner", startDate)
-        .click();
+      cy.get(".ant-picker-panel").contains(".ant-picker-cell-inner", startDate).click();
 
-      cy.get(".ant-picker-panel")
-        .contains(".ant-picker-cell-inner", endDate)
-        .click();
+      cy.get(".ant-picker-panel").contains(".ant-picker-cell-inner", endDate).click();
     };
 
     beforeEach(() => {
@@ -442,10 +446,10 @@ describe("Parameter", () => {
     });
 
     afterEach(() => {
-      cy.clock().then(clock => clock.restore());
+      cy.clock().then((clock) => clock.restore());
     });
 
-    it("updates the results after selecting a date range", function() {
+    it("updates the results after selecting a date range", function () {
       selectCalendarDateRange("15", "20");
 
       cy.getByTestId("ParameterApplyButton").click();
@@ -457,12 +461,10 @@ describe("Parameter", () => {
       );
     });
 
-    it("allows picking a dynamic date range", function() {
+    it("allows picking a dynamic date range", function () {
       cy.getByTestId("DynamicButton").click();
 
-      cy.getByTestId("DynamicButtonMenu")
-        .contains("Last month")
-        .click();
+      cy.getByTestId("DynamicButtonMenu").contains("Last month").click();
 
       cy.getByTestId("ParameterApplyButton").click();
 
@@ -479,15 +481,10 @@ describe("Parameter", () => {
   });
 
   describe("Apply Changes", () => {
-    const expectAppliedChanges = apply => {
-      cy.getByTestId("ParameterName-test-parameter-1")
-        .find("input")
-        .as("Input")
-        .type("Redash");
+    const expectAppliedChanges = (apply) => {
+      cy.getByTestId("ParameterName-test-parameter-1").find("input").as("Input").type("Redash");
 
-      cy.getByTestId("ParameterName-test-parameter-2")
-        .find("input")
-        .type("Redash");
+      cy.getByTestId("ParameterName-test-parameter-2").find("input").type("Redash");
 
       cy.location("search").should("not.contain", "Redash");
 
@@ -523,10 +520,7 @@ describe("Parameter", () => {
     it("shows and hides according to parameter dirty state", () => {
       cy.getByTestId("ParameterApplyButton").should("not.be", "visible");
 
-      cy.getByTestId("ParameterName-test-parameter-1")
-        .find("input")
-        .as("Param")
-        .type("Redash");
+      cy.getByTestId("ParameterName-test-parameter-1").find("input").as("Param").type("Redash");
 
       cy.getByTestId("ParameterApplyButton").should("be.visible");
 
@@ -536,21 +530,13 @@ describe("Parameter", () => {
     });
 
     it("updates dirty counter", () => {
-      cy.getByTestId("ParameterName-test-parameter-1")
-        .find("input")
-        .type("Redash");
+      cy.getByTestId("ParameterName-test-parameter-1").find("input").type("Redash");
 
-      cy.getByTestId("ParameterApplyButton")
-        .find(".ant-badge-count p.current")
-        .should("contain", "1");
+      cy.getByTestId("ParameterApplyButton").find(".ant-badge-count p.current").should("contain", "1");
 
-      cy.getByTestId("ParameterName-test-parameter-2")
-        .find("input")
-        .type("Redash");
+      cy.getByTestId("ParameterName-test-parameter-2").find("input").type("Redash");
 
-      cy.getByTestId("ParameterApplyButton")
-        .find(".ant-badge-count p.current")
-        .should("contain", "2");
+      cy.getByTestId("ParameterApplyButton").find(".ant-badge-count p.current").should("contain", "2");
     });
 
     it('applies changes from "Apply Changes" button', () => {
@@ -560,16 +546,13 @@ describe("Parameter", () => {
     });
 
     it('applies changes from "alt+enter" keyboard shortcut', () => {
-      expectAppliedChanges(input => {
+      expectAppliedChanges((input) => {
         input.type("{alt}{enter}");
       });
     });
 
     it('disables "Execute" button', () => {
-      cy.getByTestId("ParameterName-test-parameter-1")
-        .find("input")
-        .as("Input")
-        .type("Redash");
+      cy.getByTestId("ParameterName-test-parameter-1").find("input").as("Input").type("Redash");
       cy.getByTestId("ExecuteButton").should("be.disabled");
 
       cy.get("@Input").clear();
@@ -594,15 +577,12 @@ describe("Parameter", () => {
 
       cy.createQuery(queryData, false).then(({ id }) => cy.visit(`/queries/${id}/source`));
 
-      cy.get(".parameter-block")
-        .first()
-        .invoke("width")
-        .as("paramWidth");
+      cy.get(".parameter-block").first().invoke("width").as("paramWidth");
 
       cy.get("body").type("{alt}D"); // hide schema browser
     });
 
-    it("is possible to rearrange parameters", function() {
+    it("is possible to rearrange parameters", function () {
       cy.server();
       cy.route("POST", "**/api/queries/*").as("QuerySave");
 
