@@ -22,10 +22,7 @@ function prepareVisualization(query, type, name, options) {
       cy.get("body").type("{alt}D");
 
       // do some pre-checks here to ensure that visualization was created and is visible
-      cy.getByTestId("TableVisualization")
-        .should("exist")
-        .find("table")
-        .should("exist");
+      cy.getByTestId("TableVisualization").should("exist").find("table").should("exist");
 
       return cy.then(() => ({ queryId, visualizationId }));
     });
@@ -53,7 +50,7 @@ describe("Table", () => {
   });
 
   describe("Sorting data", () => {
-    beforeEach(function() {
+    beforeEach(function () {
       const { query, config } = MultiColumnSort;
       prepareVisualization(query, "TABLE", "Sort data", config).then(({ queryId, visualizationId }) => {
         this.queryId = queryId;
@@ -61,39 +58,22 @@ describe("Table", () => {
       });
     });
 
-    it("sorts data by a single column", function() {
-      cy.getByTestId("TableVisualization")
-        .find("table th")
-        .contains("c")
-        .should("exist")
-        .click();
+    it("sorts data by a single column", function () {
+      cy.getByTestId("TableVisualization").find("table th").contains("c").should("exist").click();
       cy.percySnapshot("Visualizations - Table (Single-column sort)", { widths: [viewportWidth] });
     });
 
-    it("sorts data by a multiple columns", function() {
-      cy.getByTestId("TableVisualization")
-        .find("table th")
-        .contains("a")
-        .should("exist")
-        .click();
+    it("sorts data by a multiple columns", function () {
+      cy.getByTestId("TableVisualization").find("table th").contains("a").should("exist").click();
 
       cy.get("body").type("{shift}", { release: false });
-      cy.getByTestId("TableVisualization")
-        .find("table th")
-        .contains("b")
-        .should("exist")
-        .click();
+      cy.getByTestId("TableVisualization").find("table th").contains("b").should("exist").click();
 
       cy.percySnapshot("Visualizations - Table (Multi-column sort)", { widths: [viewportWidth] });
     });
 
-    it("sorts data in reverse order", function() {
-      cy.getByTestId("TableVisualization")
-        .find("table th")
-        .contains("c")
-        .should("exist")
-        .click()
-        .click();
+    it("sorts data in reverse order", function () {
+      cy.getByTestId("TableVisualization").find("table th").contains("c").should("exist").click().click();
       cy.percySnapshot("Visualizations - Table (Single-column reverse sort)", { widths: [viewportWidth] });
     });
   });
@@ -101,10 +81,7 @@ describe("Table", () => {
   it("searches in multiple columns", () => {
     const { query, config } = SearchInData;
     prepareVisualization(query, "TABLE", "Search", config).then(({ visualizationId }) => {
-      cy.getByTestId("TableVisualization")
-        .find("table input")
-        .should("exist")
-        .type("test");
+      cy.getByTestId("TableVisualization").find("table input").should("exist").type("test");
       cy.percySnapshot("Visualizations - Table (Search in data)", { widths: [viewportWidth] });
     });
   });
