@@ -14,7 +14,10 @@ from sqlalchemy_utils.types.encrypted.encrypted_type import FernetEngine
 from redash import settings
 from redash.utils.configuration import ConfigurationContainer
 from redash.models.base import key_type
-from redash.models.types import EncryptedConfiguration
+from redash.models.types import (
+    EncryptedConfiguration,
+    Configuration,
+)
 
 
 # revision identifiers, used by Alembic.
@@ -42,14 +45,7 @@ def upgrade():
                 )
             ),
         ),
-        sa.Column(
-            "options", 
-            ConfigurationContainer.as_mutable(
-                EncryptedConfiguration(
-                    sa.Text, settings.DATASOURCE_SECRET_KEY, FernetEngine
-                )
-            ),
-        ),
+        sa.Column("options", ConfigurationContainer.as_mutable(Configuration)),
     )
 
     conn = op.get_bind()
