@@ -11,12 +11,12 @@ from redash.query_runner import (
 from redash.utils import json_loads
 
 try:
-    from azure.kusto.data.exceptions import KustoServiceError
-    from azure.kusto.data.request import (
+    from azure.kusto.data import (
         ClientRequestProperties,
         KustoClient,
         KustoConnectionStringBuilder,
     )
+    from azure.kusto.data.exceptions import KustoServiceError
 
     enabled = True
 except ImportError:
@@ -37,6 +37,34 @@ TYPES_MAP = {
 }
 
 
+def _get_data_scanned(kusto_response):
+    try:
+        metadata_table = next(
+            (table for table in kusto_response.tables if table.table_name == "QueryCompletionInformation"),
+            None,
+        )
+
+        if metadata_table:
+            resource_usage_json = next(
+                (row["Payload"] for row in metadata_table.rows if row["EventTypeName"] == "QueryResourceConsumption"),
+                "{}",
+            )
+            resource_usage = json_loads(resource_usage_json).get("resource_usage", {})
+
+        data_scanned = (
+            resource_usage["cache"]["shards"]["cold"]["hitbytes"]
+            + resource_usage["cache"]["shards"]["cold"]["missbytes"]
+            + resource_usage["cache"]["shards"]["hot"]["hitbytes"]
+            + resource_usage["cache"]["shards"]["hot"]["missbytes"]
+            + resource_usage["cache"]["shards"]["bypassbytes"]
+        )
+
+    except Exception:
+        data_scanned = 0
+
+    return int(data_scanned)
+
+
 class AzureKusto(BaseQueryRunner):
     should_annotate_query = False
     noop_query = "let noop = datatable (Noop:string)[1]; noop"
@@ -44,8 +72,6 @@ class AzureKusto(BaseQueryRunner):
     def __init__(self, configuration):
         super(AzureKusto, self).__init__(configuration)
         self.syntax = "custom"
-        self.client_request_properties = ClientRequestProperties()
-        self.client_request_properties.application = "redash"
 
     @classmethod
     def configuration_schema(cls):
@@ -60,12 +86,14 @@ class AzureKusto(BaseQueryRunner):
                 },
                 "azure_ad_tenant_id": {"type": "string", "title": "Azure AD Tenant Id"},
                 "database": {"type": "string"},
+                "msi": {"type": "boolean", "title": "Use Managed Service Identity"},
+                "user_msi": {
+                    "type": "string",
+                    "title": "User-assigned managed identity client ID",
+                },
             },
             "required": [
                 "cluster",
-                "azure_ad_client_id",
-                "azure_ad_client_secret",
-                "azure_ad_tenant_id",
                 "database",
             ],
             "order": [
@@ -91,18 +119,48 @@ class AzureKusto(BaseQueryRunner):
         return "Azure Data Explorer (Kusto)"
 
     def run_query(self, query, user):
-        kcsb = KustoConnectionStringBuilder.with_aad_application_key_authentication(
-            connection_string=self.configuration["cluster"],
-            aad_app_id=self.configuration["azure_ad_client_id"],
-            app_key=self.configuration["azure_ad_client_secret"],
-            authority_id=self.configuration["azure_ad_tenant_id"],
-        )
+        cluster = self.configuration["cluster"]
+        msi = self.configuration.get("msi", False)
+        # Managed Service Identity(MSI)
+        if msi:
+            # If user-assigned managed identity is used, the client ID must be provided
+            if self.configuration.get("user_msi"):
+                kcsb = KustoConnectionStringBuilder.with_aad_managed_service_identity_authentication(
+                    cluster,
+                    client_id=self.configuration["user_msi"],
+                )
+            else:
+                kcsb = KustoConnectionStringBuilder.with_aad_managed_service_identity_authentication(cluster)
+        # Service Principal auth
+        else:
+            aad_app_id = self.configuration.get("azure_ad_client_id")
+            app_key = self.configuration.get("azure_ad_client_secret")
+            authority_id = self.configuration.get("azure_ad_tenant_id")
+
+            if not (aad_app_id and app_key and authority_id):
+                raise ValueError(
+                    "Azure AD Client ID, Client Secret, and Tenant ID are required for Service Principal authentication."
+                )
+
+            kcsb = KustoConnectionStringBuilder.with_aad_application_key_authentication(
+                connection_string=cluster,
+                aad_app_id=aad_app_id,
+                app_key=app_key,
+                authority_id=authority_id,
+            )
 
         client = KustoClient(kcsb)
 
+        request_properties = ClientRequestProperties()
+        request_properties.application = "redash"
+
+        if user:
+            request_properties.user = user.email
+            request_properties.set_option("request_description", user.email)
+
         db = self.configuration["database"]
         try:
-            response = client.execute(db, query, self.client_request_properties)
+            response = client.execute(db, query, request_properties)
 
             result_cols = response.primary_results[0].columns
             result_rows = response.primary_results[0].rows
@@ -123,14 +181,15 @@ class AzureKusto(BaseQueryRunner):
                 rows.append(row.to_dict())
 
             error = None
-            data = {"columns": columns, "rows": rows}
+            data = {
+                "columns": columns,
+                "rows": rows,
+                "metadata": {"data_scanned": _get_data_scanned(response)},
+            }
 
         except KustoServiceError as err:
             data = None
-            try:
-                error = err.args[1][0]["error"]["@message"]
-            except (IndexError, KeyError):
-                error = err.args[1]
+            error = str(err)
 
         return data, error
 
@@ -143,7 +202,10 @@ class AzureKusto(BaseQueryRunner):
             self._handle_run_query_error(error)
 
         schema_as_json = json_loads(results["rows"][0]["DatabaseSchema"])
-        tables_list = schema_as_json["Databases"][self.configuration["database"]]["Tables"].values()
+        tables_list = [
+            *(schema_as_json["Databases"][self.configuration["database"]]["Tables"].values()),
+            *(schema_as_json["Databases"][self.configuration["database"]]["MaterializedViews"].values()),
+        ]
 
         schema = {}
 
@@ -154,7 +216,9 @@ class AzureKusto(BaseQueryRunner):
                 schema[table_name] = {"name": table_name, "columns": []}
 
             for column in table["OrderedColumns"]:
-                schema[table_name]["columns"].append(column["Name"])
+                schema[table_name]["columns"].append(
+                    {"name": column["Name"], "type": TYPES_MAP.get(column["CslType"], None)}
+                )
 
         return list(schema.values())
 
