@@ -1,3 +1,4 @@
+import json
 import logging
 from typing import Optional, Tuple
 
@@ -45,7 +46,7 @@ class ElasticSearch2(BaseHTTPQueryRunner):
         self.syntax = "json"
 
     def get_response(self, url, auth=None, http_method="get", **kwargs):
-        url = "{}{}".format(self.configuration["url"], url)
+        url = "{}{}".format(self.configuration["server"], url)
         headers = kwargs.pop("headers", {})
         headers["Accept"] = "application/json"
         return super().get_response(url, auth, http_method, headers=headers, **kwargs)
@@ -64,6 +65,7 @@ class ElasticSearch2(BaseHTTPQueryRunner):
         return data, error
 
     def _build_query(self, query: str) -> Tuple[dict, str, Optional[list]]:
+        query = json.loads(query)
         index_name = query.pop("index", "")
         result_fields = query.pop("result_fields", None)
         url = "/{}/_search".format(index_name)
