@@ -69,7 +69,7 @@ def datetime_parser(dct):
     return bson_object_hook(dct, json_options=opts)
 
 
-def parse_query_json(query):
+def parse_query_json(query: str):
     query_data = json_loads(query, object_hook=datetime_parser)
     return query_data
 
@@ -82,38 +82,64 @@ def _get_column_by_name(columns, column_name):
     return None
 
 
-def _parse_dict(dic):
+def _parse_dict(dic: dict, flatten: bool = False) -> dict:
     res = {}
-    for key, value in dic.items():
-        if isinstance(value, dict):
-            for tmp_key, tmp_value in _parse_dict(value).items():
-                new_key = "{}.{}".format(key, tmp_key)
-                res[new_key] = tmp_value
+
+    def _flatten(x, name=""):
+        if isinstance(x, dict):
+            for k, v in x.items():
+                _flatten(v, "{}.{}".format(name, k))
+        elif isinstance(x, list):
+            for idx, item in enumerate(x):
+                _flatten(item, "{}.{}".format(name, idx))
         else:
-            res[key] = value
+            res[name[1:]] = x
+
+    if flatten:
+        _flatten(dic)
+    else:
+        for key, value in dic.items():
+            if isinstance(value, dict):
+                for tmp_key, tmp_value in _parse_dict(value).items():
+                    new_key = "{}.{}".format(key, tmp_key)
+                    res[new_key] = tmp_value
+            else:
+                res[key] = value
     return res
 
 
-def parse_results(results):
+def parse_results(results: list, flatten: bool = False) -> list:
     rows = []
     columns = []
 
     for row in results:
         parsed_row = {}
 
-        parsed_row = _parse_dict(row)
+        parsed_row = _parse_dict(row, flatten)
         for column_name, value in parsed_row.items():
-            columns.append(
-                {
-                    "name": column_name,
-                    "friendly_name": column_name,
-                    "type": TYPES_MAP.get(type(value), TYPE_STRING),
-                }
-            )
+            if _get_column_by_name(columns, column_name) is None:
+                columns.append(
+                    {
+                        "name": column_name,
+                        "friendly_name": column_name,
+                        "type": TYPES_MAP.get(type(value), TYPE_STRING),
+                    }
+                )
 
         rows.append(parsed_row)
 
     return rows, columns
+
+
+def _sorted_fields(fields):
+    ord = {}
+    for k, v in fields.items():
+        if isinstance(v, int):
+            ord[k] = v
+        else:
+            ord[k] = len(fields)
+
+    return sorted(ord, key=ord.get)
 
 
 class MongoDB(BaseQueryRunner):
@@ -140,6 +166,14 @@ class MongoDB(BaseQueryRunner):
                     ],
                     "title": "Replica Set Read Preference",
                 },
+                "flatten": {
+                    "type": "string",
+                    "extendedEnum": [
+                        {"value": "False", "name": "False"},
+                        {"value": "True", "name": "True"},
+                    ],
+                    "title": "Flatten Results",
+                },
             },
             "secret": ["password"],
             "required": ["connectionString", "dbName"],
@@ -154,11 +188,14 @@ class MongoDB(BaseQueryRunner):
 
         self.syntax = "json"
 
-        self.db_name = self.configuration["dbName"]
+        self.db_name = self.configuration.get("dbName", "")
 
         self.is_replica_set = (
             True if "replicaSetName" in self.configuration and self.configuration["replicaSetName"] else False
         )
+
+        self.flatten = self.configuration.get("flatten", "False").upper() in ["TRUE", "YES", "ON", "1", "Y", "T"]
+        logger.debug("flatten: {}".format(self.flatten))
 
     @classmethod
     def custom_json_encoder(cls, dec, o):
@@ -278,8 +315,10 @@ class MongoDB(BaseQueryRunner):
                 if "$sort" in step:
                     sort_list = []
                     for sort_item in step["$sort"]:
-                        sort_list.append((sort_item["name"], sort_item["direction"]))
-
+                        if isinstance(sort_item, dict):
+                            sort_list.append((sort_item["name"], sort_item.get("direction", 1)))
+                        elif isinstance(sort_item, list):
+                            sort_list.append(tuple(sort_item))
                     step["$sort"] = SON(sort_list)
 
         if "fields" in query_data:
@@ -289,7 +328,10 @@ class MongoDB(BaseQueryRunner):
         if "sort" in query_data and query_data["sort"]:
             s = []
             for field_data in query_data["sort"]:
-                s.append((field_data["name"], field_data["direction"]))
+                if isinstance(field_data, dict):
+                    s.append((field_data["name"], field_data.get("direction", 1)))
+                elif isinstance(field_data, list):
+                    s.append(tuple(field_data))
 
         columns = []
         rows = []
@@ -330,16 +372,17 @@ class MongoDB(BaseQueryRunner):
 
             rows.append({"count": cursor})
         else:
-            rows, columns = parse_results(cursor)
+            rows, columns = parse_results(cursor, flatten=self.flatten)
 
         if f:
             ordered_columns = []
-            for k in sorted(f, key=f.get):
+            for k in _sorted_fields(f):
                 column = _get_column_by_name(columns, k)
                 if column:
                     ordered_columns.append(column)
 
             columns = ordered_columns
+            logger.debug("columns: {}".format(columns))
 
         if query_data.get("sortColumns"):
             reverse = query_data["sortColumns"] == "desc"
