@@ -5,6 +5,7 @@ from redash import mail, models, settings
 from redash.models import users
 from redash.query_runner import NotSupported
 from redash.tasks.worker import Queue
+from redash.version_check import run_version_check
 from redash.worker import get_job_logger, job
 
 logger = get_job_logger(__name__)
@@ -29,6 +30,27 @@ def record_event(raw_event):
             logger.exception("Failed posting to %s", hook)
 
 
+def version_check():
+    run_version_check()
+
+
+@job("default")
+def subscribe(form):
+    logger.info(
+        "Subscribing to: [security notifications=%s], [newsletter=%s]",
+        form["security_notifications"],
+        form["newsletter"],
+    )
+    data = {
+        "admin_name": form["name"],
+        "admin_email": form["email"],
+        "org_name": form["org_name"],
+        "security_notifications": form["security_notifications"],
+        "newsletter": form["newsletter"],
+    }
+    requests.post("https://version.redash.io/subscribe", json=data)
+
+
 @job("emails")
 def send_mail(to, subject, html, text):
     try:
@@ -50,7 +72,7 @@ def test_connection(data_source_id):
         return True
 
 
-@job("schemas", queue_class=Queue, at_front=True, timeout=300, ttl=90)
+@job("schemas", queue_class=Queue, at_front=True, timeout=settings.SCHEMAS_REFRESH_TIMEOUT, ttl=90)
 def get_schema(data_source_id, refresh):
     try:
         data_source = models.DataSource.get_by_id(data_source_id)
