@@ -1,6 +1,7 @@
 """
 Some test cases around the Glue catalog.
 """
+
 from unittest import TestCase
 
 import botocore
@@ -74,7 +75,9 @@ class TestGlueSchema(TestCase):
             {"DatabaseName": "test1"},
         )
         with self.stubber:
-            assert query_runner.get_schema() == [{"columns": ["row_id"], "name": "test1.jdbc_table"}]
+            assert query_runner.get_schema() == [
+                {"columns": [{"name": "row_id", "type": "int"}], "name": "test1.jdbc_table"}
+            ]
 
     def test_partitioned_table(self):
         """
@@ -123,7 +126,12 @@ class TestGlueSchema(TestCase):
             {"DatabaseName": "test1"},
         )
         with self.stubber:
-            assert query_runner.get_schema() == [{"columns": ["sk", "category"], "name": "test1.partitioned_table"}]
+            assert query_runner.get_schema() == [
+                {
+                    "columns": [{"name": "sk", "type": "int"}, {"name": "category", "type": "int"}],
+                    "name": "test1.partitioned_table",
+                }
+            ]
 
     def test_view(self):
         query_runner = Athena({"glue": True, "region": "mars-east-1"})
@@ -155,7 +163,7 @@ class TestGlueSchema(TestCase):
             {"DatabaseName": "test1"},
         )
         with self.stubber:
-            assert query_runner.get_schema() == [{"columns": ["sk"], "name": "test1.view"}]
+            assert query_runner.get_schema() == [{"columns": [{"name": "sk", "type": "int"}], "name": "test1.view"}]
 
     def test_dodgy_table_does_not_break_schema_listing(self):
         """
@@ -195,7 +203,9 @@ class TestGlueSchema(TestCase):
             {"DatabaseName": "test1"},
         )
         with self.stubber:
-            assert query_runner.get_schema() == [{"columns": ["region"], "name": "test1.csv"}]
+            assert query_runner.get_schema() == [
+                {"columns": [{"name": "region", "type": "string"}], "name": "test1.csv"}
+            ]
 
     def test_no_storage_descriptor_table(self):
         """
@@ -220,3 +230,97 @@ class TestGlueSchema(TestCase):
         )
         with self.stubber:
             assert query_runner.get_schema() == []
+
+    def test_multi_catalog_tables(self):
+        """Tables of multi-catalogs"""
+        query_runner = Athena({"glue": True, "region": "mars-east-1", "catalog_ids": "foo,bar"})
+
+        self.stubber.add_response("get_databases", {"DatabaseList": [{"Name": "test1"}]}, {"CatalogId": "foo"})
+        self.stubber.add_response(
+            "get_tables",
+            {
+                "TableList": [
+                    {
+                        "Name": "jdbc_table",
+                        "StorageDescriptor": {
+                            "Columns": [{"Name": "row_id", "Type": "int"}],
+                            "Location": "Database.Schema.Table",
+                            "Compressed": False,
+                            "NumberOfBuckets": -1,
+                            "SerdeInfo": {"Parameters": {}},
+                            "BucketColumns": [],
+                            "SortColumns": [],
+                            "Parameters": {
+                                "CrawlerSchemaDeserializerVersion": "1.0",
+                                "CrawlerSchemaSerializerVersion": "1.0",
+                                "UPDATED_BY_CRAWLER": "jdbc",
+                                "classification": "sqlserver",
+                                "compressionType": "none",
+                                "connectionName": "jdbctest",
+                                "typeOfData": "view",
+                            },
+                            "StoredAsSubDirectories": False,
+                        },
+                        "PartitionKeys": [],
+                        "TableType": "EXTERNAL_TABLE",
+                        "Parameters": {
+                            "CrawlerSchemaDeserializerVersion": "1.0",
+                            "CrawlerSchemaSerializerVersion": "1.0",
+                            "UPDATED_BY_CRAWLER": "jdbc",
+                            "classification": "sqlserver",
+                            "compressionType": "none",
+                            "connectionName": "jdbctest",
+                            "typeOfData": "view",
+                        },
+                    }
+                ]
+            },
+            {"CatalogId": "foo", "DatabaseName": "test1"},
+        )
+        self.stubber.add_response("get_databases", {"DatabaseList": [{"Name": "test2"}]}, {"CatalogId": "bar"})
+        self.stubber.add_response(
+            "get_tables",
+            {
+                "TableList": [
+                    {
+                        "Name": "jdbc_table",
+                        "StorageDescriptor": {
+                            "Columns": [{"Name": "row_id", "Type": "int"}],
+                            "Location": "Database.Schema.Table",
+                            "Compressed": False,
+                            "NumberOfBuckets": -1,
+                            "SerdeInfo": {"Parameters": {}},
+                            "BucketColumns": [],
+                            "SortColumns": [],
+                            "Parameters": {
+                                "CrawlerSchemaDeserializerVersion": "1.0",
+                                "CrawlerSchemaSerializerVersion": "1.0",
+                                "UPDATED_BY_CRAWLER": "jdbc",
+                                "classification": "sqlserver",
+                                "compressionType": "none",
+                                "connectionName": "jdbctest",
+                                "typeOfData": "view",
+                            },
+                            "StoredAsSubDirectories": False,
+                        },
+                        "PartitionKeys": [],
+                        "TableType": "EXTERNAL_TABLE",
+                        "Parameters": {
+                            "CrawlerSchemaDeserializerVersion": "1.0",
+                            "CrawlerSchemaSerializerVersion": "1.0",
+                            "UPDATED_BY_CRAWLER": "jdbc",
+                            "classification": "sqlserver",
+                            "compressionType": "none",
+                            "connectionName": "jdbctest",
+                            "typeOfData": "view",
+                        },
+                    }
+                ]
+            },
+            {"CatalogId": "bar", "DatabaseName": "test2"},
+        )
+        with self.stubber:
+            assert query_runner.get_schema() == [
+                {"columns": [{"name": "row_id", "type": "int"}], "name": "test1.jdbc_table"},
+                {"columns": [{"name": "row_id", "type": "int"}], "name": "test2.jdbc_table"},
+            ]
